// packages/adhere-util-browsersniff/src/types.ts
export interface BrowserInfo {
  /** 浏览器名称 */
  browser: string;
  /** 设备类型 */
  device: string;
  /** 渲染引擎 */
  engine: string;
  /** 语言设置 */
  language: string;
  /** 操作系统 */
  os: string;
  /** 操作系统版本 */
  osVersion: string;
  /** 浏览器版本 */
  version: string;
}
// ...
export interface IBrowserSniff {
  getInstance(): BrowserInfo;
  browser(): string;
  device(): string;
  engine(): string;
  language(): string;
  os(): string;
  osVersion(): string;
  version(): string;
  
  // 浏览器检测方法
  isBrowserSafari(): boolean;
  isBrowserChrome(): boolean;
  isBrowserIE(): boolean;
  isBrowserEdge(): boolean;
  isBrowserFirefox(): boolean;
  isBrowserFirefoxFocus(): boolean;
  isBrowserChromium(): boolean;
  isBrowserOpera(): boolean;
  isBrowserVivaldi(): boolean;
  isBrowserYandex(): boolean;
  isBrowserArora(): boolean;
  isBrowserLunascape(): boolean;
  isBrowserQupZilla(): boolean;
  isBrowserCocCoc(): boolean;
  isBrowserKindle(): boolean;
  isBrowserIceweasel(): boolean;
  isBrowserKonqueror(): boolean;
  isBrowserIceape(): boolean;
  isBrowserSeaMonkey(): boolean;
  isBrowserEpiphany(): boolean;
  isBrowser360(): boolean;
  isBrowser360EE(): boolean;
  isBrowser360SE(): boolean;
  isBrowserUC(): boolean;
  isBrowserQQBrowser(): boolean;
  isBrowserQQ(): boolean;
  isBrowserBaidu(): boolean;
  isBrowserMaxthon(): boolean;
  isBrowserSogou(): boolean;
  isBrowserLiebao(): boolean;
  isBrowser2345Explorer(): boolean;
  isBrowser115Browser(): boolean;
  isBrowserTheWorld(): boolean;
  isBrowserXiaoMi(): boolean;
  isBrowserQuark(): boolean;
  isBrowserQiyu(): boolean;
  isBrowserWechat(): boolean;
  isBrowserWechatWork(): boolean;
  isBrowserTaobao(): boolean;
  isBrowserAlipay(): boolean;
  isBrowserWeibo(): boolean;
  isBrowserDouban(): boolean;
  isBrowserSuning(): boolean;
  isBrowseriQiYi(): boolean;
  isBrowserDingTalk(): boolean;
  isBrowserHuawei(): boolean;
  isBrowserVivo(): boolean;
  isBrowserNightly(): boolean;
  
  // 设备检测方法
  isDevicePC(): boolean;
  isDeviceMobile(): boolean;
  isDeviceTablet(): boolean;
  
  // 引擎检测方法
  isEngineWebKit(): boolean;
  isEngineTrident(): boolean;
  isEngineGecko(): boolean;
  isEnginePresto(): boolean;
  isEngineKHTML(): boolean;
  isEngineBlink(): boolean;
  isEngineEdgeHTML(): boolean;
  
  // 操作系统检测方法
  iSOSWindows(): boolean;
  iSOSLinux(): boolean;
  iSOSMaxOS(): boolean;
  iSOSAndroid(): boolean;
  iSOSHarmonyOS(): boolean;
  iSOSUbuntu(): boolean;
  iSOSFreeBSD(): boolean;
  iSOSDebian(): boolean;
  iSOSiOS(): boolean;
  iSOSWindowsPhone(): boolean;
  iSOSBlackBerry(): boolean;
  iSOSMeeGo(): boolean;
  iSOSSymbian(): boolean;
  iSOSChromeOS(): boolean;
  iSOSWebOS(): boolean;
}